requirements.txt
fastapi
uvicorn
gdown
python-magic-bin
mysql-connector-python
pydantic
python-dotenv
